[utils] Move Apple Silicon W^X macros to mono-codeman.h
Need to use them both in mini/ and metadata/
